TASK2 rename src to out_of_stock_etl for convenience
