+++ tests/test_handle_product.py
@@ -1,7 +1,7 @@
 import uuid
 from typing import Final
 
-from dags.src.handler import ProductHandler
+from dags.out_of_stock_elt.handler import ProductHandler
 
 _INGESTION_TIMESTAMP: Final = "2020-12-12T12-12-12"
 _REQUESTED_DATE: Final = "2021-01-02"
@@ -16,7 +16,7 @@ def test_handle_product(
         assert_handled_product_paths,
         create_expected_target_path
 ):
-    mocked_uuid = mocker.patch('src.handler.uuid')
+    mocked_uuid = mocker.patch('out_of_stock_elt.handler.uuid')
     mocked_uuid.uuid4.return_value = _UUID4
 
     handler = ProductHandler(_INGESTION_TIMESTAMP, str(tmpdir))
